move to IPatch class EarlyStoryVisibleTimeoutPatch
