+++ src/CommunityPatch/CommunityPatchSubModule.cs
@@ -5,8 +5,6 @@ using System.Reflection;
 using System.Threading;
 using HarmonyLib;
 using JetBrains.Annotations;
-using StoryMode;
-using StoryMode.StoryModePhases;
 using TaleWorlds.CampaignSystem;
 using TaleWorlds.Core;
 using TaleWorlds.Library;
@@ -19,9 +17,6 @@ namespace CommunityPatch {
 
   [PublicAPI]
   public partial class CommunityPatchSubModule : MBSubModuleBase {
-    // may need to update this if vanilla's value changes.
-    // vanilla has it in StoryMode.Behaviors.FirstPhaseCampaignBehavior.FirstPhaseTimeLimitInYears
-    private const int FirstPhaseTimeLimitInYears = 10;
 
     internal static CommunityPatchSubModule Current
       => Module.CurrentModule.SubModules.OfType<CommunityPatchSubModule>().FirstOrDefault();
@@ -189,71 +184,8 @@ namespace CommunityPatch {
       base.OnCampaignStart(game, starterObject);
     }
 
-    private static bool IsFirstStoryPhase() {
-      return FirstPhase.Instance != null
-               && SecondPhase.Instance == null;
-    }
-
-    private void InitStoryQuestsTimeoutVisibility(Game game) {
-      if (!(game.GameType is StoryMode.CampaignStoryMode))
-        return;
-
-      CampaignEvents.OnQuestStartedEvent.AddNonSerializedListener(
-        this,
-        new Action<QuestBase>(OnQuestStarted)
-      );
-      foreach (QuestBase quest
-               in Campaign.Current.QuestManager.Quests.ToList<QuestBase>()) {
-        SetStoryVisibleTimeoutIfNeeded(quest);
-      }
-    }
-
-    private static void SetStoryVisibleTimeoutIfNeeded(QuestBase quest) {
-      if (!IsFirstStoryPhase() || !quest.IsSpecialQuest || !quest.IsOngoing)
-        return;
-
-      // set visible timeout to be when vanilla would have (silently) timed out the quest, 
-      // minus a day to make sure the quest doesn't somehow timeout due to vanilla behavior 
-      // before this.
-      CampaignTime newDueTime = FirstPhase.Instance.FirstPhaseStartTime
-                                  + CampaignTime.Years(FirstPhaseTimeLimitInYears)
-                                  - CampaignTime.Days(1);
-      if (quest.QuestDueTime != newDueTime) {
-        quest.ChangeQuestDueTime(newDueTime);
-        ShowNotification(new TextObject("{=!}Quest time remaining was updated."), "event:/ui/notification/quest_update");
-      }
-    }
-
-    private static void ShowNotification(TextObject message, string soundEventPath = "") {
-      InformationManager.AddQuickInformation(message, 0, null, soundEventPath);
-    }
-
-    class NextHourlyTickListener {
-      private Action<QuestBase> _func;
-      private QuestBase _quest;
-
-      public NextHourlyTickListener(Action<QuestBase> func, QuestBase quest) {
-        _func = func;
-        _quest = quest;
-        CampaignEvents.HourlyTickEvent.AddNonSerializedListener(this, OnHourlyTick);
-      }
-
-      private void OnHourlyTick() {
-        _func.Invoke(_quest);
-        CampaignEvents.HourlyTickEvent.ClearListeners(this);
-      }
-    };
-
-    private void OnQuestStarted(QuestBase quest) {
-      // defer our updates until some time has passed, because they depend on whether 
-      // FirstPhase or SecondPhase is active, and story Phase information is only 
-      // updated after all OnQuestStarted handlers have fired.
-      new NextHourlyTickListener(new Action<QuestBase>(SetStoryVisibleTimeoutIfNeeded), quest);
-    }
-
     public override void OnGameInitializationFinished(Game game) {
       ApplyPatches(game);
-      InitStoryQuestsTimeoutVisibility(game);
 
       base.OnGameInitializationFinished(game);
     }
